+++ src/Common/utils.ts
@@ -1,129 +1,79 @@
 import { createHash } from 'crypto'
 
-import * as addressCodec from 'ripple-address-codec'
-
-/**
- * A prefix applied when hashing a signed transaction blob.
- *
- * @see https://xrpl.org/basic-data-types.html#hashes
- */
-const signedTransactionPrefixHex = '54584E00'
-
-/**
- * A simple property bag which contains components of a classic address.
- * Components contained in this object are neither sanitized or validated.
- */
-export interface ClassicAddress {
-  /** A classic address. */
-  address: string
-
-  /** An optional tag. */
-  tag?: number
-
-  /** A boolean indicating whether this address is for use on a test network. */
-  test: boolean
-}
+import XrpUtils from '../XRP/xrp-utils'
 
 const utils = {
   /**
    * Validate that the given string is a valid address for the XRP Ledger.
-   *
    * This function returns true for both X-addresses and classic addresses.
-   * @see https://xrpaddress.info/
+   *
+   * @see {@link https://xrpaddress.info/|XRPL X-Address Format}.
+   *
+   * @deprecated Use the method provided by XrpUtils instead.
    *
    * @param address - An address to check.
+   *
    * @returns True if the address is valid, otherwise false.
    */
-  isValidAddress(address: string): boolean {
-    return (
-      addressCodec.isValidClassicAddress(address) ||
-      addressCodec.isValidXAddress(address)
-    )
-  },
+  isValidAddress: XrpUtils.isValidAddress.bind(XrpUtils),
 
   /**
    * Validate whether the given string is a valid X-address for the XRP Ledger.
    *
-   * @see https://xrpaddress.info/
+   * @see {@link https://xrpaddress.info/|XRPL X-Address Format}.
+   *
+   * @deprecated Use the method provided by XrpUtils instead.
    *
    * @param address - An address to check.
    * @returns True if the address is a valid X-address, otherwise false.
    */
-  isValidXAddress(address: string): boolean {
-    return addressCodec.isValidXAddress(address)
-  },
+  isValidXAddress: XrpUtils.isValidXAddress.bind(XrpUtils),
 
   /**
    * Validate whether the given string is a valid classic address for the XRP Ledger.
    *
-   * @see https://xrpaddress.info/
+   * @see {@link https://xrpaddress.info/|XRPL X-Address Format}.
+   *
+   * @deprecated Use the method provided by XrpUtils instead.
    *
    * @param address - An address to check.
    * @returns True if the address is a valid classic address, otherwise false.
    */
-  isValidClassicAddress(address: string): boolean {
-    return addressCodec.isValidClassicAddress(address)
-  },
+  isValidClassicAddress: XrpUtils.isValidClassicAddress.bind(XrpUtils),
 
   /**
    * Encode the given classic address and tag into an x-address.
    *
-   * @see https://xrpaddress.info/
+   * @see {@link https://xrpaddress.info/|XRPL X-Address Format}.
+   *
+   * @deprecated Use the method provided by XrpUtils instead.
    *
    * @param classicAddress - A classic address to encode.
    * @param tag - An optional tag to encode.
    * @param test - Whether the address is for use on a test network, defaults to `false`.
    * @returns A new x-address if inputs were valid, otherwise undefined.
    */
-  encodeXAddress(
-    classicAddress: string,
-    tag: number | undefined,
-    test = false,
-  ): string | undefined {
-    if (!addressCodec.isValidClassicAddress(classicAddress)) {
-      return undefined
-    }
-
-    // Xpring Common JS's API takes a string|undefined while the underlying address library wants string|false.
-    const shimTagParameter = tag ?? false
-    return addressCodec.classicAddressToXAddress(
-      classicAddress,
-      shimTagParameter,
-      test,
-    )
-  },
+  encodeXAddress: XrpUtils.encodeXAddress.bind(XrpUtils),
 
   /**
    * Decode a `ClassicAddress` from a given x-address.
    *
-   * @see https://xrpaddress.info/
+   * @see {@link https://xrpaddress.info/|XRPL X-Address Format}.
+   *
+   * @deprecated Use the method provided by XrpUtils instead.
    *
    * @param xAddress - The xAddress to decode.
-   * @returns A `ClassicAddress`
+   * @returns A `ClassicAddress`.
    */
-  decodeXAddress(xAddress: string): ClassicAddress | undefined {
-    if (!addressCodec.isValidXAddress(xAddress)) {
-      return undefined
-    }
-
-    const shimClassicAddress = addressCodec.xAddressToClassicAddress(xAddress)
-    return {
-      address: shimClassicAddress.classicAddress,
-      tag:
-        shimClassicAddress.tag === false ? undefined : shimClassicAddress.tag,
-      test: shimClassicAddress.test,
-    }
-  },
+  decodeXAddress: XrpUtils.decodeXAddress.bind(XrpUtils),
 
   /**
    * Convert the given byte array to a hexadecimal string.
    *
-   * @param bytes - An array of bytes
+   * @param bytes - An array of bytes.
    * @returns An encoded hexadecimal string.
    */
-  toHex(bytes: Uint8Array): string {
-    return Buffer.from(bytes).toString('hex').toUpperCase()
-  },
+  toHex: XrpUtils.toHex.bind(XrpUtils),
 
   /**
    * Convert the given hexadecimal string to a byte array.
@@ -138,28 +88,22 @@ const utils = {
   /**
    * Convert the given transaction blob to a transaction hash.
    *
+   * @deprecated Use the method provided by XrpUtils instead.
+   *
    * @param transactionBlobHex - A hexadecimal encoded transaction blob.
    * @returns A hex encoded hash if the input was valid, otherwise undefined.
    */
   transactionBlobToTransactionHash(
     transactionBlobHex: string,
   ): string | undefined {
-    if (!this.isHex(transactionBlobHex)) {
-      return undefined
-    }
-
-    const prefixedTransactionBlob = this.toBytes(
-      signedTransactionPrefixHex + transactionBlobHex,
-    )
-    const hash = this.sha512Half(prefixedTransactionBlob)
-    return this.toHex(hash)
+    return XrpUtils.transactionBlobToTransactionHash(transactionBlobHex)
   },
 
   /**
    * Check if the given string is valid hex.
    *
    * @param input - The input to check.
-   * @returns true if the input is valid hex, otherwise false.
+   * @returns True if the input is valid hex, otherwise false.
    */
   isHex(input: string): boolean {
     const hexRegEx = /(?:[0-9]|[a-f])/gimu
@@ -169,7 +113,7 @@ const utils = {
   /**
    * Compute the SHA512 half hash of the given bytes.
    *
-   * @param input - The input to hash.
+   * @param bytes - The array of bytes to hash.
    * @returns The hash of the input.
    */
   sha512Half(bytes: Uint8Array): Uint8Array {
@@ -180,6 +124,12 @@ const utils = {
     return hash.slice(0, hash.length / 2)
   },
 
+  /**
+   * Convert a buffer to a hex string.
+   *
+   * @param buffer - The buffer to convert.
+   * @returns A string representing the buffer.
+   */
   hexFromBuffer(buffer: Buffer): string {
     return buffer.toString('hex').toUpperCase()
   },