+++ src/PayID/pay-id-components.ts
@@ -1,6 +1,8 @@
 // Disable multiple classes to accommodate the switch to idiomatic style naming.
 // TODO(keefertaylor): Remove this when migration is complete.
-/* eslint-disable  max-classes-per-file */
+/* eslint-disable  max-classes-per-file --
+ * We only export two classes in this file because one is deprecated and will eventually be removed.
+ */
 
 /**
  * A class which encapsulates components of a Pay ID.
@@ -25,7 +27,7 @@ export default class PayIdComponents {
 /**
  * A class which encapsulates components of a Pay ID.
  *
- * @deprecated Please use the idiomatically named `PayIdComponents` class instead.
+ * @deprecated Use the idiomatically named `PayIdComponents` class instead.
  */
 export class PayIDComponents {
   public readonly host: string